requirements.txt
streamlit
openai-whisper
torch       
ffmpeg-python
noisereduce 
librosa 
soundfile 
numpy
matplotlib